+++ addOns/soap/src/main/javahelp/org/zaproxy/zap/extension/soap/resources/help_ja_JP/helpset_ja_JP.hs
@@ -9,7 +9,7 @@
 
   <view>
     <name>TOC</name>
-    <label>コンテンツ</label>
+    <label>目次</label>
     <type>org.zaproxy.zap.extension.help.ZapTocView</type>
     <data>toc.xml</data>
   </view>